dt-bindings: pinctrl: Add Nuvoton WPCM450
This binding is heavily based on the one for NPCM7xx, because the
hardware is similar. One notable difference is that there are no
sub-nodes for GPIO banks, because the GPIO registers are arranged
differently.

Certain pins support blink patterns in hardware. This is currently not
modelled in the DT binding.

Signed-off-by: Jonathan Neuschäfer <j.neuschaefer@gmx.net>
---

Should this binding be merged with the one for the NPCM7xx pinctrl/gpio
controller?


Should the following be possible?

	pinctrl@... {
		spi {
			groups = "gspi", "sspi";
			function = "gpio";
		};
	};

i.e. explicitly switching groups to GPIO.