ethernet: Add Nuvoton EMC (Ethernet MAC Controller) driver
This driver is based on the Nuvoton W90P910 Ethernet driver, which was
removed in commit 00d2fbf73d550 ("net: remove w90p910-ether driver").
The intention is to use it with other Nuvoton SoCs, specifically the
WPCM450 and NPCM7xx BMCs.

Compared to the state of the driver when it was removed, this patch
makes a few changes:

- The name is changed to nuvoton-emc to reflect its more general nature
- The driver can now be configured via devicetree
- Register descriptions were added (based on the Winbond W90N745CD datasheet)
- Various refactoring and modernization


Known issues:
- NC/SI support is untested

---

Known problems:
- MAC address configuration is fake
- avoid magic names like "rxbd"