ARM: compressed: Pass the actual output length to the decompressor
ZSTD writes outside of the space that is necessary for the uncompressed
data, when it is told it has unlimited output length. To fix this, pass
the actual output length (the length of the uncompressed kernel) to the
decompressor.

The uncompressed length is already stored as a little endian 32-bit
constant before the input_data_end symbol. This has been the case since
commit bc22c17e12c13 ("bzip2/lzma: library support for gzip, bzip2 and
lzma decompression"), which has been merged in 2.6.30.

Signed-off-by: Jonathan Neuschäfer <j.neuschaefer@gmx.net>
Reviewed-by: Linus Walleij <linus.walleij@linaro.org>
Reviewed-by: Florian Fainelli <f.fainelli@gmail.com>
Reviewed-by: Tony Lindgren <tony@atomide.com>
---

v2:
- Document when the appended size appeared
- Add R-b tags