[WIP] ARM: dts: wpcm450: Remove clock-output-names from reference clock node
A previous version of the as-yet unmerged clk-wpcm450 driver[1] used the
output name, but in the current iteration it doesn't rely on it anymore.
Thus it can be removed from the devicetree.

[1]: Added in "clk: wpcm450: Add Nuvoton WPCM450 clock/reset controller driver"

Fixes: 362e8be2ec04a6 ("ARM: dts: wpcm450: Add clock controller node")
Signed-off-by: Jonathan Neuschäfer <j.neuschaefer@gmx.net>
---

TODO: rename node to clock-ref or something, because it isn't
necessarily 48MHz, just by convention

v11:
- Specify since when clock-output-names is unnecessary

v10:
- no changes

v9:
- New patch