powerpc: mpc83xx: Switch to of_platform_default_populate
Quoting from drivers/of/platform.c:

> of_platform_populate() - [...]
> Similar to of_platform_bus_probe(), this function walks the device
> tree and creates devices from nodes.  It differs in that it follows
> the modern convention of requiring all device nodes to have a
> 'compatible' property, and it is suitable for creating devices which
> are children of the root node (of_platform_bus_probe will only create
> children of the root which are selected by the @matches argument).

This is useful for new board ports because it means that the C code does
not have to anticipate every node that is placed directly under the root.

As a consequence, the of_bus_ids list can be replaced by the default list;
this is what of_platform_default_populate does.

NOTE: This change breaks the ABI between device tree files and kernel in
such a way that a bus marked only with device_type = "soc" will not be
populated anymore. In-tree, the only affected DT is asp834x-redboot.dts,
which is adjusted in the previous commit.

Signed-off-by: J. Neuschäfer <j.ne@posteo.net>
---

V2:
- document ABI break
- remove "gianfar" compatible string (handled in a separate patch)
- switch to of_platform_default_populate (makes the code even shorter)