powerpc: boot: Enable FIT image generation
The Flat Image Tree (FIT) format combines a kernel, a set of
devicetrees, and optionally additional resources into a single file that
can be loaded by a bootloader such as U-Boot. Generating a FIT image as
part of the kernel build reduces the need for additional build scripts,
and produces a single boot image without falling back to one of the
many legacy methods implemented in arch/powerpc/boot/Makefile, which
would require additional changes for arch/powerpc/boot for each board.

Compression is handled by make_fit.py and the cmd_fit logic in
scripts/Makefile.lib. The kernel and device trees are gzip-compressed
by default.

Signed-off-by: J. Neuschäfer <j.ne@posteo.net>
---

V2:
- Document in the commit message how compression is handled.

V1:
Checkpatch warns:

  arch/powerpc/boot/Makefile:364: WARNING: Possible repeated word: 'Image'

for this line:

  image-y += Image image.fit

This is bogus and can be ignored.