powerpc: MPC83xx cleanup and LANCOM NWAPP2 board
This patchset adds a devicetree for the LANCOM NWAPP2 board based on the
MPC8314E platform, and slightly modernizes the MPC83xx platform support
in the process.

This board is nominally end-of-life, but available to hobbyists.
A U-Boot port is also in the making.


To: Christophe Leroy <christophe.leroy@csgroup.eu>
Signed-off-by: J. Neuschäfer <j.ne@posteo.net>

---
Changes in v2:
- dropped the following patches, because they have already been merged:
  [11/19] dt-bindings: gpio: fsl,qoriq-gpio: Add compatible string fsl,mpc8314-gpio
  [13/19] gpio: mpc8xxx: Add MPC8314 support
- Link to v1: https://lore.kernel.org/r/20250102-mpc83xx-v1-0-86f78ba2a7af@posteo.net



--- b4-submit-tracking ---
{
  "series": {
    "revision": 2,
    "change-id": "20241231-mpc83xx-17c898a49441",
    "prefixes": [],
    "history": {
      "v1": [
        "20250102-mpc83xx-v1-0-86f78ba2a7af@posteo.net"
      ]
    }
  }
}