powerpc: dts: asp834x-redboot: Add missing compatible to SoC bus
In preparation for switching the MPC83xx platform support to
of_platform_[default_]populate, I noticed that the asp834x-redboot.dts
is the only MPC83xx devicetree that does not list a compatible string
for the SoC bus, but rather only device_type = "soc", which doesn't work
with of_platform_populate.

This patch adds compatible = "simple-bus", which prepares
asp834x-redboot.dts for the DT ABI-breaking change of switching to
of_platform_populate.

Signed-off-by: J. Neuschäfer <j.ne@posteo.net>
---

V2:
- new patch