net: gianfar: Probe bus under ethernet node
A gianfar ethernet controller specified in a device tree contains a
subnode that represents the MDIO bus, which in turn may contain PHYs.

On Freescale/NXP platforms such as MPC83xx, the MDIO node was so far
probed because { .compatible = "gianfar" } was specified in the matches
argument of of_platform_populate.

Prepare to switch to of_platform_default_populate on MPC83xx by ensuring
in the gianfar driver that subnodes of a gianfar ethernet controller
node are probed.

Signed-off-by: J. Neuschäfer <j.ne@posteo.net>
---

V2:
- new patch