ARM: dts: amlogic: Add TCU Fernsehfee 3.0
Fernsehfee[1] ("TV fairy") 3.0 is a set-top box with HDMI input and
output ports. It originally ran Android 4.4 and a Linux 3.10 kernel.

The following features are tested and known to work:

- Ethernet
- Power LED (switching between green and red)
- Power button
- eMMC
- SD Card
- USB
- Wifi

The following features are untested or not working:

- HDMI input and output
- Infrared remote control input and output

[1]: https://fernsehfee.de/ (German), https://telefairy.com/ (English)

Reviewed-by: Neil Armstrong <neil.armstrong@linaro.org>
Signed-off-by: J. Neuschäfer <j.ne@posteo.net>
---
This patch depends on a new binding, which has entered linux-next
through usb-next: "dt-bindings: net: wireless: Add Realtek RTL8188ETV USB WiFi"

  https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/log/?qt=grep&q=RTL8188ETV

V2:
- Fix bias-pull-up property
- Remove unused sd_card_slot label, apply Neil Armstrong's R-b tag
- Move vdd-supply to wifi module, instead of using vbus-supply
- Remove regulator-1v8, because measurements did not show evidence that
  the eMMC chip actually gets 1.8V for Vccq.